requirements.txt
flask
flask_socketio
fer
logging
opencv-python
deepface
tf_keras
ultralytics
pyrealsense2
moviepy
imageio-ffmpeg
pyserial
streamlit 
azure-cognitiveservices-speech 
openai
streamlit_chat
soundfile
sounddevice